#54 fix another problem when summing experiments

+++ _pytadbit/experiment.py
@@ -207,7 +207,7 @@ class Experiment(object):
             other.norm = norm2
         xpr.crm = self.crm
         if not xpr.size:
-            xpr.size = sqrt(len(xpr.norm[0]))
+            xpr.size = int(sqrt(len(xpr.norm[0])))
         
         def __merge(own, fgn):
             "internal function to merge descriptions"
